refactor: items関連処理をuseTextItems.tsにまとめた #74
